requirements.txt
# API Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Document Processing
python-docx==1.1.2
python-pptx==1.0.2
pdfplumber==0.11.4
openpyxl==3.1.5

# Search
rank-bm25==0.2.2

# LLM - Google Gemini (free tier)
google-genai==1.14.0

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.3.4
httpx==0.28.1
